feat: Add ob-fsharp & adjust lookup
Add a VERY rudimentary org-babel exporter for F#. It literally just runs
fsi on a file and is likely to break. After AoC, this will be cleaned up
and likely PRd into org-contrib.

Seems having `+docsets` on lookup makes everything slow to a crawl on
macOS, and I never used it anyway.